Introducing TimeoutTimer
TimeoutTimer is a utility class for setting timeouts that may change while
it's armed. I could have used redy-made classes for this, but I decided to
write my own because I want to add some features to this in the future that
most likely other timer classes won't have (named and time-limited timer
suspension).