Make the suite runner throw TestFailureError when a test fails
This will make it possible for the one that uses the suite runner to know
when to show the error and when to suppess it. Test failures don't need to
be shown usually, because they will already have been reported by the
reporters. Other errors, however, need to be properly reported.