Emit breadcrumbs before hooks and each test
This makes it easier to do nice reporting with breadcrumbs. Most of the time,
it is not ideal to leave a breadcrumb of a previous phase of a test as the
"latest" breadcrumb. This fixes that and makes things more clear (while still
keeping the possiblity for reporters to do something different with this because
of the systemGenerated flag)